Question 4 (b)
You are provided with balanced binary tree with the target value k. return x number of values that are closest to the
given target k. provide solution in O(n)
Note: You have only one set of unique values x in binary search tree that are closest to the target. // Example tree:
        BinaryTree tree = new BinaryTree();
        tree.insert(4);
        tree.insert(2);
        tree.insert(5);
        tree.insert(1);
        tree.insert(3);

        double target(k) = 3.8;
        int x = 2;

        Find Output: 3, 4