Question 5 (b)
Assume you were hired to create an application for an ISP, and there are n network devices, such as routers,
that are linked together to provide internet access to users. You are given a 2D array that represents network
connections between these network devices. write an algorithm to return impacted network devices, If there is
a power outage on a certain device, these impacted device list assist you notify linked consumers that there is a
power outage and it will take some time to rectify an issue.

Input: edges= {{0,1},{0,2},{1,3},{1,6},{2,4},{4,6},{4,5},{5,7}}
Target Device (On which power Failure occurred): 4
Output (Impacted Device List) = {5,7}